safari-534.53-branch/LayoutTests/svg/animations/animate-dynamic-update-attributeName-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Test behavior on dynamic-update of attributeName

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
